Renamed module to donut to follow PEP8
