+++ .gitignore
@@ -1,4 +1,4 @@
-Donut/config.py
+donut/config.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
