+++ donut/__init__.py
@@ -6,16 +6,16 @@ import traceback
 import httplib
 import datetime
 
-from Donut import constants
+from donut import constants
 
-from Donut.modules import auth
+from donut.modules import auth
 
 app = flask.Flask(__name__)
 app.debug = False
 
 # Get app config, if we're not testing on travis.
 if 'TRAVIS' not in os.environ:
-  app.config.from_object('Donut.config')
+  app.config.from_object('donut.config')
 
 # Maximum file upload size, in bytes.
 app.config['MAX_CONTENT_LENGTH'] = constants.MAX_CONTENT_LENGTH
@@ -58,4 +58,4 @@ def internal_server_error(error):
   return flask.render_template("500.html"), httplib.INTERNAL_SERVER_ERROR
 
 # After initialization, import the routes.
-from Donut import routes
+from donut import routes
