+++ donut/modules/auth/__init__.py
@@ -1,4 +1,4 @@
 import flask
 blueprint = flask.Blueprint('auth', __name__, template_folder='templates')
 
-import Donut.modules.auth.routes
+import donut.modules.auth.routes
